+++ NEWS.md
@@ -1,5 +1,13 @@
-# betaselectr 0.0.1.1
+# betaselectr 0.0.1.2
 
 - Added `lm_betaselect()` and related
   methods and helper functions.
   (0.0.1.1)
+
+- Added parallel processing support to
+  the internal function
+  `find_all_products()`. (0.0.1.2)
+
+- For `lav_betaselect()`, added an
+  option to skip finding product
+  terms. (0.0.1.2)